Update MainSketch.ino
I added looping variables towards the top of the program and in each function/method pertaining to sensor input, I added a few lines that upload the data captured serially to the Pi. The value of time is something we need to update so that it gets the correct time from the hasp. Or depending if we want to we can start our own timer, then capture the data from the temperature and other sensors do a bit of cross-referencing then determine when to start the experiment or look for interesting data in etc. The value x is another variable we should change to what we want. It would be how many seconds the observation should last. We could instead have everything running constantly the only problem with that is the power consumption.