SerialInput
This is just a bit of simple code to gather data from the Arduino on the pi and store it in a .txt file. There are some parts of the code that still need adjusting such as the port location and the the .txt file location. There is a part inside the code that decodes the input from Unicode(which it automatically inputs as) and the keyboard interrupt is something we can work out or I believe is unnecessary and could be cut off by apart of a batch command sequence.